+++ src/main/java/leetcode/tree/AverageOfLevels.java
@@ -6,7 +6,9 @@ import java.util.List;
 import java.util.Queue;
 
 /*
- * Given a non-empty binary tree, return the average value of the nodes on each level in the form of an array.
+ * LEETCODE 637
+ * Given a non-empty binary tree, return the average value of the nodes 
+ * on each level in the form of an array.
  * Example 1:
  * Input:
  *     3
@@ -15,12 +17,13 @@ import java.util.Queue;
  *     /  \
  *    15   7
  * Output: [3, 14.5, 11]
- * Explanation: The average value of nodes on level 0 is 3,  on level 1 is 14.5, and on level 2 is 11. Hence return [3, 14.5, 11].
+ * Explanation: The average value of nodes on level 0 is 3,  on level 1 is 14.5, 
+ * and on level 2 is 11. Hence return [3, 14.5, 11].
  * Note: The range of node's value is in the range of 32-bit signed integer.
  * 
  * Company: Facebook
  * Difficulty: easy
- * Similar Questions: 102(Binary Tree Level Order Traversal), 107(102(Binary Tree Level Order Traversal II)
+ * Similar Questions: 102(Binary Tree Level Order Traversal), 107(Binary Tree Level Order Traversal II)
  */
 public class AverageOfLevels {
     public List<Double> averageOfLevels(TreeNode root) {
@@ -31,7 +34,7 @@ public class AverageOfLevels {
         Queue<TreeNode> queue = new LinkedList<>();
         queue.offer(root);
         while (!queue.isEmpty()) {
-            long sum = 0;
+            double sum = 0.0;
             int size = queue.size();
             for (int i = 0; i < size; i++) {
                 TreeNode node = queue.poll();
@@ -41,7 +44,7 @@ public class AverageOfLevels {
                     queue.offer(node.right);
                 sum += node.val;
             }
-            result.add(Double.valueOf(sum) / size);
+            result.add(sum / size);
         }
 
         return result;
